+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Added
+
+- Parse single environment variable using `env.get()`
+
 ## [0.0.3] - 2026-03-03
 
 ### Added